python/requirements.txt
flask
flask_cors
mysql-connector-python
sentence-transformers
torch
torchvision
numpy
Pillow
python-dotenv
deepface 
opencv-python
tf-keras
Sastrawi
spacy
langdetect
fuzzywuzzy
xgboost
python-Levenshtein
APScheduler